refactor: merge common and cement parameter reconciliation to cement only
